% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOLL-G.R
\name{BOLLG}
\alias{BOLLG}
\alias{pbollg}
\alias{dbollg}
\alias{qbollg}
\alias{rbollg}
\alias{hbollg}
\title{The beta Odd log-logistic family of distributions (BOLL-G)}
\usage{
pbollg(x, alpha = 1, a = 1, b = 1, G = pnorm, ...)

dbollg(x, alpha = 1, a = 1, b = 1, G = pnorm, ...)

qbollg(q, alpha = 1, a = 1, b = 1, G = pnorm, ...)

rbollg(n, alpha = 1, a = 1, b = 1, G = pnorm, ...)

hbollg(x, alpha = 1, a = 1, b = 1, G = pnorm, ...)
}
\arguments{
\item{x, q}{A numeric/quantiles    vector.}

\item{alpha}{non-negative parameter.}

\item{a}{non-negative parameter.}

\item{b}{non-negative parameter.}

\item{G}{A baseline continuous cdf.}

\item{...}{The baseline cdf parameters.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\value{
\code{pbollg} gives the distribution function,
 \code{dbollg} gives the density,
 \code{qbollg} gives the quantile function,
 \code{hbollg} gives the hazard function and
 \code{rbollg} generates random variables from the The beta Odd log-logistic family of
 distributions (BOLL-G) for baseline cdf G.
}
\description{
Distribution function, density, quantile function, hazard
 function and random generation for The beta Odd log-logistic family
 of distributions (BOLL-G) with baseline cdf G.
}
\examples{
x <- seq(0, 1, length.out = 21)
pbollg(x)
pbollg(x, alpha = 2, a = 2, b = 2, G = pbeta, shape1 = 1, shape2 = 2)
dbollg(x, alpha = 2, a = 2, b = 2, G = pbeta, shape1 = 1, shape2 = 2)
curve(dbollg, -3, 3)
qbollg(x, alpha = 2, a = 2, b = 2, G = pbeta, shape1 = 1, shape2 = 2)
n <- 10
rbollg(n, alpha = 2, a = 2, b = 2, G = pbeta, shape1 = 1, shape2 = 2)
hbollg(x, alpha = 2, a = 2, b = 2, G = pbeta, shape1 = 1, shape2 = 2)
curve(hbollg, -3, 3)
}
\references{
Cordeiro, G. M., Alizadeh, M., Tahir, M. H., Mansoor, M., Bourguignon, M., Hamedani, G. G. (2016). The beta odd log-logistic generalized family of distributions. Hacettepe Journal of Mathematics and Statistics, 45(4), 1175-1202.
}
