% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KwOLL-G.R
\name{KwOLLG}
\alias{KwOLLG}
\alias{pkwollg}
\alias{dkwollg}
\alias{qkwollg}
\alias{rkwollg}
\alias{hkwollg}
\title{Kumaraswamy Odd log-logistic family of distributions (KwOLL-G)}
\usage{
pkwollg(x, alpha = 1, a = 1, b = 1, G = pnorm, ...)

dkwollg(x, alpha = 1, a = 1, b = 1, G = pnorm, ...)

qkwollg(q, alpha = 1, a = 1, b = 1, G = pnorm, ...)

rkwollg(n, alpha = 1, a = 1, b = 1, G = pnorm, ...)

hkwollg(x, alpha = 1, a = 1, b = 1, G = pnorm, ...)
}
\arguments{
\item{x, q}{A numeric/quantiles    vector.}

\item{alpha}{non-negative parameter.}

\item{a}{non-negative parameter.}

\item{b}{non-negative parameter.}

\item{G}{A baseline continuous cdf.}

\item{...}{The baseline cdf parameters.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\value{
\code{pkwollg} gives the distribution function,
 \code{dkwollg} gives the density,
 \code{qkwollg} gives the quantile function,
 \code{hkwollg} gives the hazard function and
 \code{rkwollg} generates random variables from the Kumaraswamy Odd log-logistic family of
 distributions (KwOLL-G) for baseline cdf G.
}
\description{
Distribution function, density, quantile function, hazard
 function and random generation for Kumaraswamy Odd log-logistic family
 of distributions (KwOLL-G) with baseline cdf G.
}
\examples{
x <- seq(0, 1, length.out = 21)
pkwollg(x)
pkwollg(x, alpha = 2, a = 1, b = 1, G = pbeta, shape1 = 1, shape2 = 2)
dkwollg(x, alpha = 2, a = 1, b = 1, G = pbeta, shape1 = 1, shape2 = 2)
curve(dkwollg, -3, 3)
qkwollg(x, alpha = 2, a = 1, b = 1, G = pbeta, shape1 = 1, shape2 = 2)
n <- 10
rkwollg(n, alpha = 2, a = 1, b = 1, G = pbeta, shape1 = 1, shape2 = 2)
hkwollg(x, alpha = 2, a = 1, b = 1, G = pbeta, shape1 = 1, shape2 = 2)
curve(hkwollg, -3, 3)
}
\references{
Alizadeh, M., Emadi, M., Doostparast, M., Cordeiro, G. M., Ortega, E. M., Pescim, R. R. (2015). A new family of distributions: the Kumaraswamy odd log-logistic, properties and applications. Hacettepe Journal of Mathematics and Statistics, 44(6), 1491-1512.
}
