% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOOLL-G.R
\name{MOOLLG}
\alias{MOOLLG}
\alias{pmoollg}
\alias{dmoollg}
\alias{qmoollg}
\alias{rmoollg}
\alias{hmoollg}
\title{Marshal-Olkin Odd log-logistic family of distributions (MOOLL-G)}
\usage{
pmoollg(x, alpha = 1, beta = 1, G = pnorm, ...)

dmoollg(x, alpha = 1, beta = 1, G = pnorm, ...)

qmoollg(q, alpha = 1, beta = 1, G = pnorm, ...)

rmoollg(n, alpha = 1, beta = 1, G = pnorm, ...)

hmoollg(x, alpha = 1, beta = 1, G = pnorm, ...)
}
\arguments{
\item{x, q}{A numeric/quantiles    vector.}

\item{alpha}{non-negative parameters.}

\item{beta}{non-negative parameters.}

\item{G}{A baseline continuous cdf.}

\item{...}{The baseline cdf parameters.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\value{
\code{pmoollg} gives the distribution function,
 \code{dmoollg} gives the density,
 \code{qmoollg} gives the quantile function,
 \code{hmoollg} gives the hazard function and
 \code{rmoollg} generates random variables from the Marshal-Olkin Odd log-logistic family of
 distributions (MOOLL-G) for baseline cdf G.
}
\description{
Distribution function, density, quantile function, hazard
 function and random generation for Marshal-Olkin Odd log-logistic family
 of distributions (MOOLL-G) with baseline cdf G.
}
\examples{
x <- seq(0, 1, length.out = 21)
pmoollg(x)
pmoollg(x, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
dmoollg(x, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
curve(dmoollg, -3, 3)
qmoollg(x, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
n <- 10
rmoollg(n, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
hmoollg(x, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
curve(hmoollg, -3, 3)
}
\references{
Gleaton, J. U., Lynch, J. D. (2010). Extended generalized loglogistic families of lifetime distributions with an application. J. Probab. Stat.Sci, 8(1), 1-17.
}
