% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBOLL-G.R
\name{RBOLLG}
\alias{RBOLLG}
\alias{prbollg}
\alias{drbollg}
\alias{qrbollg}
\alias{rrbollg}
\alias{hrbollg}
\title{The Ristic-Balakrishnan  Odd log-logistic family of distributions (RBOLL-G)}
\usage{
prbollg(x, alpha = 1, beta = 1, G = pnorm, ...)

drbollg(x, alpha = 1, beta = 1, G = pnorm, ...)

qrbollg(q, alpha = 1, beta = 1, G = pnorm, ...)

rrbollg(n, alpha = 1, beta = 1, G = pnorm, ...)

hrbollg(x, alpha = 1, beta = 1, G = pnorm, ...)
}
\arguments{
\item{x, q}{A numeric/quantiles    vector.}

\item{alpha}{non-negative parameter.}

\item{beta}{non-negative parameter..}

\item{G}{A baseline continuous cdf.}

\item{...}{The baseline cdf parameters.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\value{
\code{prbollg} gives the distribution function,
 \code{drbollg} gives the density,
 \code{qrbollg} gives the quantile function,
 \code{hrbollg} gives the hazard function and
 \code{rrbollg} generates random variables from the The Ristic-Balakrishnan  Odd log-logistic family of
 distributions (RBOLL-G) for baseline cdf G.
}
\description{
Distribution function, density, quantile function, hazard
 function and random generation for The Ristic-Balakrishnan  Odd log-logistic family
 of distributions (RBOLL-G) with baseline cdf G.
}
\examples{
x <- seq(0, 1, length.out = 21)
prbollg(x)
prbollg(x, alpha = 2, beta = 1, G = pbeta, shape1 = 1, shape2 = 2)
drbollg(x, alpha = 2, beta = 1, G = pbeta, shape1 = 1, shape2 = 2)
curve(drbollg, -3, 3)
qrbollg(x, alpha = 2, beta = 1, G = pbeta, shape1 = 1, shape2 = 2)
n <- 10
rrbollg(n, alpha = 2, beta = 1, G = pbeta, shape1 = 1, shape2 = 2)
hrbollg(x, alpha = 2, beta = 1, G = pbeta, shape1 = 1, shape2 = 2)
curve(hrbollg, -3, 3)
}
\references{
Esmaeili, H., Lak, F., Altun, E. (2020). The Ristic-Balakrishnan odd log-logistic family of distributions: Properties and Applications. Statistics, Optimization Information Computing, 8(1), 17-35.
}
