% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZBOLL-G.R
\name{ZBOLLG}
\alias{ZBOLLG}
\alias{pzbollg}
\alias{dzbollg}
\alias{qzbollg}
\alias{rzbollg}
\alias{hzbollg}
\title{The Zografos-Balakrishnan  Odd log-logistic family of distributions (ZBOLL-G)}
\usage{
pzbollg(x, alpha = 1, beta = 1, G = pnorm, ...)

dzbollg(x, alpha = 1, beta = 1, G = pnorm, ...)

qzbollg(q, alpha = 1, beta = 1, G = pnorm, ...)

rzbollg(n, alpha = 1, beta = 1, G = pnorm, ...)

hzbollg(x, alpha = 1, beta = 1, G = pnorm, ...)
}
\arguments{
\item{x, q}{A numeric/quantiles    vector.}

\item{alpha}{non-negative parameter.}

\item{beta}{non-negative parameter..}

\item{G}{A baseline continuous cdf.}

\item{...}{The baseline cdf parameters.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\value{
\code{pzbollg} gives the distribution function,
 \code{dzbollg} gives the density,
 \code{qzbollg} gives the quantile function,
 \code{hzbollg} gives the hazard function and
 \code{rzbollg} generates random variables from the The Zografos-Balakrishnan  Odd log-logistic family of
 distributions (ZBOLL-G) for baseline cdf G.
}
\description{
Distribution function, density, quantile function, hazard
 function and random generation for The Zografos-Balakrishnan  Odd log-logistic family
 of distributions (ZBOLL-G) with baseline cdf G.
}
\examples{
x <- seq(0, 1, length.out = 21)
pzbollg(x)
pzbollg(x, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
dzbollg(x, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
curve(dzbollg, -3, 3)
qzbollg(x, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
n <- 10
rzbollg(n, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
hzbollg(x, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
curve(hzbollg, -3, 3)
}
\references{
Cordeiro, G. M., Alizadeh, M., Ortega, E. M., Serrano, L. H. V. (2016). The Zografos-Balakrishnan odd log-logistic family of distributions: Properties and Applications. Hacettepe Journal of Mathematics and Statistics, 45(6), 1781-1803. .
}
