% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOLL-G.R
\name{EOLLG}
\alias{EOLLG}
\alias{peollg}
\alias{deollg}
\alias{qeollg}
\alias{reollg}
\alias{heollg}
\title{Exponentiated Odd log-logistic family of distributions (EOLL-G)}
\usage{
peollg(x, alpha = 1, beta = 1, G = pnorm, ...)

deollg(x, alpha = 1, beta = 1, G = pnorm, ...)

qeollg(q, alpha = 1, beta = 1, G = pnorm, ...)

reollg(n, alpha = 1, beta = 1, G = pnorm, ...)

heollg(x, alpha = 1, beta = 1, G = pnorm, ...)
}
\arguments{
\item{x, q}{A numeric/quantiles    vector.}

\item{alpha}{non-negative parameters.}

\item{beta}{non-negative parameters.}

\item{G}{A baseline continuous cdf.}

\item{...}{The baseline cdf parameters.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\value{
\code{peollg} gives the distribution function,
 \code{deollg} gives the density,
 \code{qeollg} gives the quantile function,
 \code{heollg} gives the hazard function and
 \code{reollg} generates random variables from the Exponentiated Odd log-logistic family of
 distributions (EOLL-G) for baseline cdf G.
}
\description{
Distribution function, density, quantile function, hazard
 function and random generation for Exponentiated Odd log-logistic family
 of distributions (EOLL-G) with baseline cdf G.
}
\examples{
x <- seq(0, 1, length.out = 21)
peollg(x)
peollg(x, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
deollg(x, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
curve(deollg, -3, 3)
qeollg(x, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
n <- 10
reollg(n, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
heollg(x, alpha = 2, beta = 1, G = pbeta, shape1 = 1, shape2 = 2)
curve(heollg, -3, 3)
}
\references{
Cordeiro , G . M . , Alizadeh , M . , Ozel , G . , Hosseini , B . , Ortega , E . M . M . , Altun , E . ( 2017 ) . The generalized odd log - logistic family of distributions : properties , regression models and applications . Journal of Statistical Computation and Simulation , 87 ( 5 ) , 908 - 932 .
}
