% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOLL-G.R
\name{GOLLG}
\alias{GOLLG}
\alias{pgollg}
\alias{dgollg}
\alias{qgollg}
\alias{rgollg}
\alias{hgollg}
\title{Generalized Odd log-logistic family of distributions (GOLL-G)}
\usage{
pgollg(x, alpha = 1, beta = 1, G = pnorm, ...)

dgollg(x, alpha = 1, beta = 1, G = pnorm, ...)

qgollg(q, alpha = 1, beta = 1, G = pnorm, ...)

rgollg(n, alpha = 1, beta = 1, G = pnorm, ...)

hgollg(x, alpha = 1, beta = 1, G = pnorm, ...)
}
\arguments{
\item{x, q}{A numeric/quantiles    vector.}

\item{alpha}{non-negative parameter.}

\item{beta}{non-negative parameter.}

\item{G}{A baseline continuous cdf.}

\item{...}{The baseline cdf parameters.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\value{
\code{pgollg} gives the distribution function,
 \code{dgollg} gives the density,
 \code{qgollg} gives the quantile function,
 \code{hgollg} gives the hazard function and
 \code{rgollg} generates random variables from the Generalized Odd log-logistic family of
 distributions (GOLL-G) for baseline cdf G.
}
\description{
Distribution function, density, quantile function, hazard
 function and random generation for Generalized Odd log-logistic family
 of distributions (GOLL-G) with baseline cdf G.
}
\examples{
x <- seq(0, 1, length.out = 21)
pgollg(x)
pgollg(x, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
dgollg(x, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
curve(dgollg, -3, 3)
qgollg(x, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
n <- 10
rgollg(n, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
hgollg(x, alpha = 2, beta = 1, G = pbeta, shape1 = 1, shape2 = 2)
curve(hgollg, -3, 3)
}
\references{
Cordeiro, G.M., Alizadeh, M., Ozel, G., Hosseini, B., Ortega, E.M.M., Altun, E. (2017). The generalized odd log-logistic family of distributions : properties, regression models and applications. Journal of Statistical Computation and Simulation ,87(5),908-932.
}
