% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOLL-G.R
\name{NOLLG}
\alias{NOLLG}
\alias{pnollg}
\alias{dnollg}
\alias{qnollg}
\alias{rnollg}
\alias{hnollg}
\title{New Odd log-logistic family of distributions (NOLL-G)}
\usage{
pnollg(x, alpha = 1, beta = 1, G = pnorm, ...)

dnollg(x, alpha = 1, beta = 1, G = pnorm, ...)

qnollg(q, alpha = 1, beta = 1, G = pnorm, ...)

rnollg(n, alpha = 1, beta = 1, G = pnorm, ...)

hnollg(x, alpha = 1, beta = 1, G = pnorm, ...)
}
\arguments{
\item{x, q}{A numeric/quantiles    vector.}

\item{alpha}{non-negative parameters.}

\item{beta}{non-negative parameters.}

\item{G}{A baseline continuous cdf.}

\item{...}{The baseline cdf parameters.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\value{
\code{pnollg} gives the distribution function,
 \code{dnollg} gives the density,
 \code{qnollg} gives the quantile function,
 \code{hnollg} gives the hazard function and
 \code{rnollg} generates random variables from the New Odd log-logistic family of
 distributions (NOLL-G) for baseline cdf G.
}
\description{
Distribution function, density, quantile function, hazard
 function and random generation for New Odd log-logistic family
 of distributions (NOLL-G) with baseline cdf G.
}
\examples{
x <- seq(0, 1, length.out = 21)
pnollg(x)
pnollg(x, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
dnollg(x, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
curve(dnollg, -3, 3)
qnollg(x, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
n <- 10
rnollg(n, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
hnollg(x, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
curve(hnollg, -3, 3)
}
\references{
Alizadeh, M., Altun, E., Ozel, G., Afshari, M., Eftekharian, A. (2019). A new odd log-logistic lindley distribution with properties and applications. Sankhya A, 81(2), 323-346.
}
