% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOLL-G.R, R/oll-g.R
\name{OLLG}
\alias{OLLG}
\alias{hanollg}
\alias{pollg}
\alias{dollg}
\alias{qollg}
\alias{rollg}
\alias{hollg}
\title{Odd log-logistic family of distributions (OLL-G)}
\usage{
hanollg(x, alpha = 1, beta = 1, G = pnorm, ...)

pollg(x, alpha = 1, G = pnorm, ...)

dollg(x, alpha = 1, G = pnorm, ...)

qollg(q, alpha = 1, G = pnorm, ...)

rollg(n, alpha = 1, G = pnorm, ...)

hollg(x, alpha = 1, G = pnorm, ...)
}
\arguments{
\item{x, q}{A numeric/quantiles    vector.}

\item{alpha}{non-negative parameters.}

\item{G}{A baseline continuous cdf.}

\item{...}{The baseline cdf parameters.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\value{
\code{pollg} gives the distribution function,
 \code{dollg} gives the density,
 \code{qollg} gives the quantile function,
 \code{hollg} gives the hazard function and
 \code{rollg} generates random variables from the Odd log-logistic family of
 distributions (OLL-G) for baseline cdf G.
}
\description{
Distribution function, density, quantile function, hazard
 function and random generation for Odd log-logistic family
 of distributions (OLL-G) with baseline cdf G.
}
\examples{
hanollg(x, alpha = 2, beta = 2, G = pbeta, shape1 = 1, shape2 = 2)
curve(hanollg, -3, 3)
x <- seq(0, 1, length.out = 21)
pollg(x)
pollg(x, alpha = 2, G = pbeta, shape1 = 1, shape2 = 2)

dollg(x, alpha = 2, G = pbeta, shape1 = 1, shape2 = 2)
curve(dollg, -3, 3)
qollg(x, alpha = 2, G = pbeta, shape1 = 1, shape2 = 2)
n <- 10
rollg(n, alpha = 2, G = pbeta, shape1 = 1, shape2 = 2)
hollg(x, alpha = 2, G = pbeta, shape1 = 1, shape2 = 2)
curve(hollg, -3, 3)
}
